# src/swagger.yaml
openapi: 3.0.0
info:
  title: Tutor App API
  contact: {}
  version: '1.0'
servers:
- url: https://ffprac-team2-back.onrender.com/api/v1/
  variables: {}
paths:
...
  /tutors:
    post:
      tags:
      - Tutors
      summary: Add New Tutor
      operationId: AddNewTutor
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AddNewTutorRequest'
              - example:
                  grades:
                  - '6'
                  - '7'
                  - '8'
                  avatar: ''
                  yearsOfExperience: 10
                  education: Harvard
                  availability:
                  - Monday
                  - Tuesday
                  - Wednesday
                  subject:
                  - math
                  - science
                  MathSubject:
                  - Algebra
                  - Geometry
                  Science:
                  - Chemistry
                  - Biology
                  ForeignLanguages:
                  - Spanish
                  - Chinese
                  about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
            example:
              grades:
              - '6'
              - '7'
              - '8'
              avatar: ''
              yearsOfExperience: 10
              education: Harvard
              availability:
              - Monday
              - Tuesday
              - Wednesday
              subject:
              - math
              - science
              MathSubject:
              - Algebra
              - Geometry
              Science:
              - Chemistry
              - Biology
              ForeignLanguages:
              - Spanish
              - Chinese
              about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
        required: true
      responses:
        '201':
          description: CREATED
          headers: {}
        '401':
          description: UNAUTHORIZED
          headers: {}
        '403':
          description: FORBIDDEN
          headers: {}
        '500':
          description: BAD_REQUEST
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Tutors
      summary: Get All Tutors
      operationId: GetAllTutors
      parameters:
      - in: query
        name: subjects
        description: Comma-separated list of subjects
        schema:
          type: string
      - in: query
        name: grades
        description: Comma-separated list of grades
        schema:
          type: string
      responses:
        '200':
          description: OK
          headers: {}
        '500':
          description: BAD_REQUEST
          headers: {}
      deprecated: false
      security: []
...
  /tutors/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The tutor ID
          
    get:
      tags:
      - Tutors
      summary: Get 1 Tutor By ID
      operationId: Get1TutorByID
      responses:
        '200':
          description: OK
          headers: {}
        '401':
          description: UNAUTHORIZED
          headers: {}
        '403':
          description: FORBIDDEN
          headers: {}
        '500':
          description: BAD_REQUEST
          headers: {}
      deprecated: false
      security: []
    patch:
      tags:
      - Tutors
      summary: Update Tutor
      operationId: UpdateTutor
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateTutorRequest'
              - example:
                  yearsOfExperience: 10
                  education: Harvard
                  grades:
                  - '11'
                  - '12'
                  avatar: ''
                  availability:
                  - Monday
                  - Tuesday
                  - Wednesday
                  subject:
                  - math
                  - science
                  MathSubject:
                  - Algebra
                  - Geometry
                  Science:
                  - Chemistry
                  - Biology
                  ForeignLanguages:
                  - Spanish
                  - Chinese
                  about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
            example:
              yearsOfExperience: 10
              education: Harvard
              grades:
              - '11'
              - '12'
              avatar: ''
              availability:
              - Monday
              - Tuesday
              - Wednesday
              subject:
              - math
              - science
              MathSubject:
              - Algebra
              - Geometry
              Science:
              - Chemistry
              - Biology
              ForeignLanguages:
              - Spanish
              - Chinese
              about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
        required: true
      responses:
        '200':
          description: OK
          headers: {}
        '401':
          description: UNAUTHORIZED
          headers: {}
        '403':
          description: FORBIDDEN
          headers: {}
        '500':
          description: BAD_REQUEST
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Tutors
      summary: Delete Tutor
      operationId: DeleteTutor
      responses:
        '200':
          description: OK
          headers: {}
        '401':
          description: UNAUTHORIZED
          headers: {}
        '403':
          description: FORBIDDEN
          headers: {}
        '500':
          description: BAD_REQUEST
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    AddNewTutorRequest:
      title: AddNewTutorRequest
      required:
      - grades
      - avatar
      - yearsOfExperience
      - education
      - availability
      - subject
      - MathSubject
      - Science
      - ForeignLanguages
      - about
      type: object
      properties:
        grades:
          type: array
          items:
            type: string
          description: ''
        avatar:
          type: string
        yearsOfExperience:
          type: integer
          format: int32
        education:
          type: string
        availability:
          type: array
          items:
            type: string
          description: ''
        subject:
          type: array
          items:
            type: string
          description: ''
        MathSubject:
          type: array
          items:
            type: string
          description: ''
        Science:
          type: array
          items:
            type: string
          description: ''
        ForeignLanguages:
          type: array
          items:
            type: string
          description: ''
        about:
          type: string
      example:
        grades:
        - '6'
        - '7'
        - '8'
        avatar: ''
        yearsOfExperience: 10
        education: Harvard
        availability:
        - Monday
        - Tuesday
        - Wednesday
        subject:
        - math
        - science
        MathSubject:
        - Algebra
        - Geometry
        - Algebra
        Science:
        - Chemistry
        - Biology
        ForeignLanguages:
        - Spanish
        - Chinese
        about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
    UpdateTutorRequest:
      title: UpdateTutorRequest
      required:
      - yearsOfExperience
      - education
      - grades
      - avatar
      - availability
      - subject
      - MathSubject
      - Science
      - ForeignLanguages
      - about
      type: object
      properties:
        yearsOfExperience:
          type: integer
          format: int32
        education:
          type: string
        grades:
          type: array
          items:
            type: string
          description: ''
        avatar:
          type: string
        availability:
          type: array
          items:
            type: string
          description: ''
        subject:
          type: array
          items:
            type: string
          description: ''
        MathSubject:
          type: array
          items:
            type: string
          description: ''
        Science:
          type: array
          items:
            type: string
          description: ''
        ForeignLanguages:
          type: array
          items:
            type: string
          description: ''
        about:
          type: string
      example:
        yearsOfExperience: 10
        education: Harvard
        grades:
        - '11'
        - '12'
        avatar: ''
        availability:
        - Monday
        - Tuesday
        - Wednesday
        subject:
        - math
        - science
        MathSubject:
        - Algebra
        - Geometry
        Science:
        - Chemistry
        - Biology
        ForeignLanguages:
        - Spanish
        - Chinese
        about: I'm a passionate and experienced Mathematics tutor with more than 10 years of teaching experience. I have a Bachelor's degree in Mathematics and a Master's degree in Education, both from NYU.
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...